Add cells to check train set has full range of feature/target values
The range of values for the features and target in the train set, for
some algorithms at least, determines the range of validity of the model:
for which feature/target values the model can be applied. So added cells
to check that the extreme variable values of the dataset are present in
the train set. Likewise ensured all categorical variable values feature
in the train set.